src/apps/requirements.txt
dash==2.5.1
dash_bootstrap_components==1.1.0
plotly==5.8.0
flask-caching==2.0.1
flask-login==0.6.2
ldap3==2.9.1
